# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LearningExperiment": {
        "properties": {
          "arms": {
            "items": {
              "$ref": "#/components/schemas/LearningExperimentArm"
            },
            "title": "Arms",
            "type": "array"
          },
          "bestArmName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bestarmname"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "feedbackCategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedbackcategory"
          },
          "feedbackDriven": {
            "title": "Feedbackdriven",
            "type": "boolean"
          },
          "feedbackReasonCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedbackreasoncode"
          },
          "hypothesis": {
            "title": "Hypothesis",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "methodologyVersion": {
            "title": "Methodologyversion",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "hypothesis",
          "status",
          "createdAt",
          "arms",
          "bestArmName",
          "feedbackDriven",
          "feedbackCategory",
          "feedbackReasonCode",
          "methodologyVersion"
        ],
        "title": "LearningExperiment",
        "type": "object"
      },
      "LearningExperimentArm": {
        "properties": {
          "accuracy": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "armName": {
            "title": "Armname",
            "type": "string"
          },
          "horizonDaysFilter": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizondaysfilter"
          },
          "modelVersion": {
            "title": "Modelversion",
            "type": "string"
          },
          "sampleCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Samplecount"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "windowLabel": {
            "title": "Windowlabel",
            "type": "string"
          }
        },
        "required": [
          "armName",
          "modelVersion",
          "windowLabel",
          "horizonDaysFilter",
          "sampleCount",
          "accuracy",
          "verdict"
        ],
        "title": "LearningExperimentArm",
        "type": "object"
      },
      "LearningHistoryEntry": {
        "description": "One event in the unified learning-history timeline. `id` is a\ncomposite `\"<source-table>:<row id>\"` string (not a single row's own\nprimary key) since this feed merges rows from several distinct,\nindependently-keyed tables -- never a fabricated shared sequence.",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "decisionReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decisionreason"
          },
          "evidenceCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidencecount"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "impact": {
            "title": "Impact",
            "type": "string"
          },
          "methodologyVersion": {
            "title": "Methodologyversion",
            "type": "string"
          },
          "modelVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelversion"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "createdAt",
          "status",
          "evidenceCount",
          "methodologyVersion",
          "impact",
          "modelVersion",
          "decisionReason"
        ],
        "title": "LearningHistoryEntry",
        "type": "object"
      },
...
      "Meta": {
        "properties": {
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "timestamp"
        ],
        "title": "Meta",
        "type": "object"
      },
...
      "SuccessEnvelope_list_LearningExperiment__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/LearningExperiment"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessEnvelope[list[LearningExperiment]]",
        "type": "object"
      },
      "SuccessEnvelope_list_LearningHistoryEntry__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/LearningHistoryEntry"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessEnvelope[list[LearningHistoryEntry]]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Market Agent M1",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/learning/experiments": {
      "get": {
        "operationId": "get_learning_experiments_endpoint_api_v1_learning_experiments_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_list_LearningExperiment__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Learning Experiments Endpoint",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/learning/history": {
      "get": {
        "operationId": "get_learning_history_endpoint_api_v1_learning_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_list_LearningHistoryEntry__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Learning History Endpoint",
        "tags": [
          "learning"
        ]
      }
    },
...
  }
}